выносит комментарии в отдельную структуру
